Создаёт контроллер и переносит логику работы компонентов в него.
